services/edi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
uvloop==0.20.0
asyncpg==0.29.0
httpx==0.27.0
pydantic==2.9.0
python-multipart==0.0.9
asyncssh==2.17.0
PyJWT==2.9.0
